Added Rules.has_ops to provide an array of players have ops at any given position.  The default implementation returns an array containing the player whose turn it is now.
Changed the signatures of Rules.ops and Rules.op? so that they now take an
optional player parameter.  If the player is not specified or nil ops and op?
behave as before.  If a player is specified the ops set only includes ops that
that player should be able to apply.  Similarly, op? only returns true if the
given op is valid for the given player.  Rules.apply does not care or check
player -- clients of the library can check if they wish.

The above two changes should allow support for simultaneous moves.

The changes and a new convention allow support for random events.  If a ops are
available for psuedo player :random, then whatever client is controlling the
game should apply one of the ops at random.  The list could be weighted by
providing the same op multiple times.  This support is highly dependant on
convention and fairly rough.

Updated all existing rules to use the new ops and op? signatures.

Added rules for Pig to demonstrate the new support for random events.

Added TrueClass and FalseClass to the list of dup exceptions in PositionStruct.