Added Rules.censor.  It creates a Rules.censor method (confusing, I know) that overwrites the specified fields with :hidden.  In this way a Position can be censored before showing it to a player (to hide cards, for example).
Added Rules.display to provide a way to indicate which Position fields should
be shown to the player.  This is to separate bookkeeping fields from the
interesting fields.  (for example, #turn is usually redundant because of
Rules.has_ops)

Updated all Rules to use display.  Also, updated Minesweeper to use censor.