Added Connect6Board as subclass of Board.  It adds tracking of "threats", which made AI easier / more efficient.
Added bots for Connect6.  NeighborsBot will probably be removed at some point
in the near future (very inefficient), but ThreatsBot plays a decent (but not
great) game.

Modified AlphaBeta search to add support for prune and order methods.  (to
prune and/or order ops lists)  Minimax will need to be updated to remain
completely compatible.