Added Rules.ignore and Rules.ignored?.
Bug fix:  moves_cache was not being used in Othello (and Makyek), because the 
instance variable @moves_cache was not being assigned to.

Added moves_cache support to Ataxx, creating a noticable speed up.  It'd be
nice to pick up some more speed though.

Added calls to .ignore in several rules subclasses.