Bug fixes:  The previous changed marked a move back to allowing Position#moves to return any object (as opposed to Strings only).  Some unit tests had to be updated to include extra #to_s calls.
One of the search tests broke because there were actually two valid return
values and the one being returned changed when the order of #moves changed in
the previous update to TicTacToe.

Added Game#moves that behaves in the same way Game#move? does.  Fixed the
related unit test.  It was broken long before the last release (though the code
was not *necessarily* broken).