Bug fix:  Hearts.pass_before_deal was being modified (it's frozen, but the value of :directions is not).
Changed method signature for #moves, #apply!.

Updated unit tests which were generally broken (not sure how they were even
passing) with regard to pass_before_deal.  (Some tests were written as if there
would be no passing phase)

Hearts / TrickTaking is generally very messed up and in need of a thorough
refactor.  This is not it.  This is simply a temporary fix / update.