Gut a lot of the complexity in Bot.
Bots used to have a setup where a single named bot could play multiple games by
defining namespaced Bot extentions like:

    class MyBot < Bot
      # We'd mostly refer to MyBot
      class TicTacToe < Bot
        # But when playing TicTacToe we'd delegate to this class
      end
    end

This scheme is sort of weird / difficult and I'm not sure we need to support
such a complicated scheme for Bots.

This also removes a couple usages of the weird nested_const stuff that it'd be
nice to remove.